Added Model for Networks Added missing Attributes to Series and Episode Added missing model für Season
