+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/models/Collection.java
@@ -1,5 +1,4 @@
-package de.fourzerofournotfound.rateyourstuff.RaYS.models;
-
+package de.fourzerofournotfound.rateyourstuff.rays.models;
 
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -7,16 +6,17 @@ import org.hibernate.annotations.ColumnDefault;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
+import java.util.Set;
 
 @Data
 @Entity
 @RequiredArgsConstructor
-@Table(name = "LoginRoles")
-public class LoginRole {
+@Table(name = "Collections")
+public class Collection {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long loginRoleId;
+    private Long collectionId;
 
     @Column(nullable = false)
     @ColumnDefault("CURRENT_TIMESTAMP()")
@@ -27,11 +27,9 @@ public class LoginRole {
     private LocalDateTime updatedAt;
 
     @ManyToOne
-    @JoinColumn(name = "loginID", referencedColumnName = "loginID")
-    private Login login;
+    @JoinColumn(name = "userId", referencedColumnName = "userid")
+    private User user;
 
-    @ManyToOne
-    @JoinColumn(name = "roleId", referencedColumnName = "roleId")
-    private Role role;
-
+    @ManyToMany
+    Set<Medium> media;
 }
