+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/models/Episode.java
@@ -1,22 +1,24 @@
-package de.fourzerofournotfound.rateyourstuff.RaYS.models;
+package de.fourzerofournotfound.rateyourstuff.rays.models;
 
-import lombok.Data;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.Set;
 
-@Data
+@Getter
+@Setter
 @Entity
 @RequiredArgsConstructor
-@Table(name = "Collections")
-public class Collection {
+@Table(name = "Episodes")
+public class Episode {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long collectionId;
+    private Long episodeId;
 
     @Column(nullable = false)
     @ColumnDefault("CURRENT_TIMESTAMP()")
@@ -26,10 +28,20 @@ public class Collection {
     @ColumnDefault("NULL ON UPDATE CURRENT_TIMESTAMP()")
     private LocalDateTime updatedAt;
 
-    @ManyToOne
-    @JoinColumn(name = "userId", referencedColumnName = "userid")
-    private User user;
+    @Column(length = 250)
+    private String episodeTitle;
 
-    @ManyToMany
-    Set<Medium> media;
+    @Column(length = 2000)
+    private String episodeDescription;
+
+    @Column
+    private Integer episodeNumber;
+
+    @Column
+    private Integer length;
+
+    @ManyToOne
+    @JoinColumn(name = "seasonId", referencedColumnName = "seasonId")
+    private Season season;
 }
+
