+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/models/Network.java
@@ -1,0 +1,25 @@
+package de.fourzerofournotfound.rateyourstuff.rays.models;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+@RequiredArgsConstructor
+@Table(name="networks")
+public class Network {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long networkId;
+
+    @Column
+    private String networkTitle;
+
+    @OneToMany (mappedBy = "network")
+    private List<Series> series;
+}
