+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/models/Person.java
@@ -1,4 +1,4 @@
-package de.fourzerofournotfound.rateyourstuff.RaYS.models;
+package de.fourzerofournotfound.rateyourstuff.rays.models;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -6,18 +6,21 @@ import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 
 import javax.persistence.*;
+import java.sql.Date;
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Getter
 @Setter
 @Entity
 @RequiredArgsConstructor
-@Table(name = "Episodes")
-public class Episode {
+@Table(name = "Persons")
+
+public class Person {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long episodeId;
+    private Long personId;
 
     @Column(nullable = false)
     @ColumnDefault("CURRENT_TIMESTAMP()")
@@ -26,5 +29,16 @@ public class Episode {
     @Column(nullable = true)
     @ColumnDefault("NULL ON UPDATE CURRENT_TIMESTAMP()")
     private LocalDateTime updatedAt;
-}
 
+    @Column (length=250)
+    private String firstName;
+
+    @Column (length=250)
+    private String lastName;
+
+    @Column
+    private Date birthDate;
+
+    @OneToMany (mappedBy = "person")
+    private List<PersonAssignment> personAssignments;
+}
