+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/models/Platform.java
@@ -1,0 +1,36 @@
+package de.fourzerofournotfound.rateyourstuff.rays.models;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.ColumnDefault;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.util.Set;
+
+@Getter
+@Setter
+@Entity
+@RequiredArgsConstructor
+@Table(name = "Platforms")
+public class Platform {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long plattformId;
+
+    @Column(nullable = false)
+    @ColumnDefault("CURRENT_TIMESTAMP()")
+    private LocalDateTime createdAt;
+
+    @Column
+    @ColumnDefault("NULL ON UPDATE CURRENT_TIMESTAMP()")
+    private LocalDateTime updatedAt;
+
+    @Column (length = 250)
+    private String platformTitel;
+
+    @ManyToMany
+    Set<Medium> media;
+
+}
