+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/repositories/MovieRepository.java
@@ -1,0 +1,8 @@
+package de.fourzerofournotfound.rateyourstuff.rays.repositories;
+
+import de.fourzerofournotfound.rateyourstuff.rays.models.Movie;
+import org.springframework.data.repository.CrudRepository;
+
+public interface MovieRepository extends CrudRepository<Movie, Long> {
+
+}
