+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/repositories/SeasonRepository.java
@@ -1,0 +1,7 @@
+package de.fourzerofournotfound.rateyourstuff.rays.repositories;
+
+import de.fourzerofournotfound.rateyourstuff.rays.models.Season;
+import org.springframework.data.repository.CrudRepository;
+
+public interface SeasonRepository extends CrudRepository<Season, Long> {
+}
