+++ src/main/java/de/fourzerofournotfound/rateyourstuff/rays/repositories/SeriesRepository.java
@@ -1,0 +1,7 @@
+package de.fourzerofournotfound.rateyourstuff.rays.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.http.HttpStatus;
+
+public interface SeriesRepository extends CrudRepository<HttpStatus.Series, Long> {
+}
